Continue Diamond scenarios
Continue to expand diamond scenarios; several tests are planned but not yet implemented.
